Rename `graphql.rs` to `api.rs`
There were similarly named types in two files.
- `graphql::Issue` and `github::Issues`
- `graphql::PullRequest` and `github::PullRequests`

Since both sets share similar contexts related to GraphQL, it was easy
to confuse them while reading the code. Renaming `graphql` to `api`
makes it clearer that these types are used to serve the server-side API.

Closes #147
