+++ src/api/pull_request.rs
@@ -9,56 +9,59 @@ use async_graphql::{
 use crate::database::{self, Database, TryFromKeyValue};
 
 #[derive(SimpleObject)]
-pub(crate) struct Issue {
+pub(crate) struct PullRequest {
     pub(crate) owner: String,
     pub(crate) repo: String,
     pub(crate) number: i32,
     pub(crate) title: String,
-    pub(crate) author: String,
+    pub(crate) assignees: Vec<String>,
+    pub(crate) reviewers: Vec<String>,
 }
 
-impl TryFromKeyValue for Issue {
+impl TryFromKeyValue for PullRequest {
     fn try_from_key_value(key: &[u8], value: &[u8]) -> anyhow::Result<Self> {
         let (owner, repo, number) = database::parse_key(key)
             .with_context(|| format!("invalid key in database: {key:02x?}"))?;
-        let (title, author, _) = bincode::deserialize::<(String, String, Option<String>)>(value)?;
-        let issue = Issue {
+        let deserialized = bincode::deserialize::<(String, Vec<String>, Vec<String>)>(value)?;
+        let (title, assignees, reviewers) = deserialized;
+        let pr = PullRequest {
             title,
-            author,
+            assignees,
+            reviewers,
             owner,
             repo,
             number: i32::try_from(number).unwrap_or(i32::MAX),
         };
-        Ok(issue)
+        Ok(pr)
     }
 }
 
-impl fmt::Display for Issue {
+impl fmt::Display for PullRequest {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         write!(f, "{}/{}#{}", self.owner, self.repo, self.number)
     }
 }
 
 #[derive(Default)]
-pub(super) struct IssueQuery;
+pub(super) struct PullRequestQuery;
 
 #[Object]
-impl IssueQuery {
-    async fn issues(
+impl PullRequestQuery {
+    async fn pull_requests(
         &self,
         ctx: &Context<'_>,
         after: Option<String>,
         before: Option<String>,
         first: Option<i32>,
         last: Option<i32>,
-    ) -> Result<Connection<String, Issue, EmptyFields, EmptyFields>> {
+    ) -> Result<Connection<String, PullRequest, EmptyFields, EmptyFields>> {
         query(
             after,
             before,
             first,
             last,
             |after, before, first, last| async move {
-                super::load_connection(ctx, Database::issues, after, before, first, last)
+                super::load_connection(ctx, Database::pull_requests, after, before, first, last)
             },
         )
         .await
@@ -67,14 +70,14 @@ impl IssueQuery {
 
 #[cfg(test)]
 mod tests {
-    use crate::{github::GitHubIssue, graphql::TestSchema};
+    use crate::{api::TestSchema, github::GitHubPullRequests};
 
     #[tokio::test]
-    async fn issues_empty() {
+    async fn pull_requests_empty() {
         let schema = TestSchema::new();
         let query = r"
         {
-            issues {
+            pullRequests {
                 edges {
                     node {
                         number
@@ -83,37 +86,34 @@ mod tests {
             }
         }";
         let res = schema.execute(query).await;
-        assert_eq!(res.data.to_string(), "{issues: {edges: []}}");
+        assert_eq!(res.data.to_string(), "{pullRequests: {edges: []}}");
     }
 
     #[tokio::test]
-    async fn issues_first() {
+    async fn pull_requests_first() {
         let schema = TestSchema::new();
-        let issues = vec![
-            GitHubIssue {
+        let pull_requests = vec![
+            GitHubPullRequests {
                 number: 1,
-                title: "issue 1".to_string(),
-                author: "author 1".to_string(),
-                closed_at: None,
+                title: "pull request 1".to_string(),
+                assignees: vec!["assignee 1".to_string()],
+                reviewers: vec!["reviewer 1".to_string()],
             },
-            GitHubIssue {
+            GitHubPullRequests {
                 number: 2,
-                title: "issue 2".to_string(),
-                author: "author 2".to_string(),
-                closed_at: None,
-            },
-            GitHubIssue {
-                number: 3,
-                title: "issue 3".to_string(),
-                author: "author 3".to_string(),
-                closed_at: None,
+                title: "pull request 2".to_string(),
+                assignees: vec!["assignee 2".to_string()],
+                reviewers: vec!["reviewer 2".to_string()],
             },
         ];
-        schema.db.insert_issues(issues, "owner", "name").unwrap();
+        schema
+            .db
+            .insert_pull_requests(pull_requests, "owner", "name")
+            .unwrap();
 
         let query = r"
         {
-            issues(first: 2) {
+            pullRequests(first: 1) {
                 edges {
                     node {
                         number
@@ -127,12 +127,12 @@ mod tests {
         let res = schema.execute(query).await;
         assert_eq!(
             res.data.to_string(),
-            "{issues: {edges: [{node: {number: 1}}, {node: {number: 2}}], pageInfo: {hasNextPage: true}}}"
+            "{pullRequests: {edges: [{node: {number: 1}}], pageInfo: {hasNextPage: true}}}"
         );
 
         let query = r"
         {
-            issues(first: 5) {
+            pullRequests(first: 5) {
                 pageInfo {
                     hasNextPage
                 }
@@ -141,38 +141,35 @@ mod tests {
         let res = schema.execute(query).await;
         assert_eq!(
             res.data.to_string(),
-            "{issues: {pageInfo: {hasNextPage: false}}}"
+            "{pullRequests: {pageInfo: {hasNextPage: false}}}"
         );
     }
 
     #[tokio::test]
-    async fn issues_last() {
+    async fn pull_requests_last() {
         let schema = TestSchema::new();
-        let issues = vec![
-            GitHubIssue {
+        let pull_requests = vec![
+            GitHubPullRequests {
                 number: 1,
-                title: "issue 1".to_string(),
-                author: "author 1".to_string(),
-                closed_at: None,
+                title: "pull request 1".to_string(),
+                assignees: vec!["assignee 1".to_string()],
+                reviewers: vec!["reviewer 1".to_string()],
             },
-            GitHubIssue {
+            GitHubPullRequests {
                 number: 2,
-                title: "issue 2".to_string(),
-                author: "author 2".to_string(),
-                closed_at: None,
-            },
-            GitHubIssue {
-                number: 3,
-                title: "issue 3".to_string(),
-                author: "author 3".to_string(),
-                closed_at: None,
+                title: "pull request 2".to_string(),
+                assignees: vec!["assignee 2".to_string()],
+                reviewers: vec!["reviewer 2".to_string()],
             },
         ];
-        schema.db.insert_issues(issues, "owner", "name").unwrap();
+        schema
+            .db
+            .insert_pull_requests(pull_requests, "owner", "name")
+            .unwrap();
 
         let query = r"
         {
-            issues(last: 2) {
+            pullRequests(last: 1) {
                 edges {
                     node {
                         number
@@ -186,12 +183,12 @@ mod tests {
         let res = schema.execute(query).await;
         assert_eq!(
             res.data.to_string(),
-            "{issues: {edges: [{node: {number: 2}}, {node: {number: 3}}], pageInfo: {hasPreviousPage: true}}}"
+            "{pullRequests: {edges: [{node: {number: 2}}], pageInfo: {hasPreviousPage: true}}}"
         );
 
         let query = r"
         {
-            issues(last: 5) {
+            pullRequests(last: 2) {
                 pageInfo {
                     hasPreviousPage
                 }
@@ -200,7 +197,7 @@ mod tests {
         let res = schema.execute(query).await;
         assert_eq!(
             res.data.to_string(),
-            "{issues: {pageInfo: {hasPreviousPage: false}}}"
+            "{pullRequests: {pageInfo: {hasPreviousPage: false}}}"
         );
     }
 }
